+++ history.md
@@ -1,4 +1,4 @@
-## Edge
+## Version 2.0.0 / 2012-04-24
 
 * Add more helpers to `wti init` a project
 * Add ability to add a source locale to a project from the command line: the first `wti addlocale xx` sets the source locale.